+++ docs/CHANGELOG.md
@@ -8,4 +8,4 @@
 - Created documentation for GitHub workflow and pull request template.
 - Added maven workspace and configuration.
 - Added Use Cases 4 and 5
-
+- Implement entities for stock and stock market